fix click events re-propogation on mobile
fixes GH-170

Previously, we did not fire a click event for touch events
on mobile unless the event target was a form field. This is
incorrect as the event was not `preventDefault`'d before.

As you can see on Lines 48-50 in 0e596c (not the commit
that introduced the bug, just a hard link reference),
the "begin scroll" handler does not preventDefault
on the event if it is a form field, so we do not need
 to re-propogate it in `synthesizeClick`.

All other tag types *do* need to synthesize the click
event on mobile however.

I ran the tests in the XCode simulator and they passed.